+++ quickstart/src/main/kotlin/jp/shiguredo/sora/quickstart/MainActivity.kt
@@ -1,6 +1,7 @@
 package jp.shiguredo.sora.quickstart
 
 import android.Manifest
+import android.annotation.SuppressLint
 import android.content.Context
 import android.graphics.Color
 import android.media.AudioManager
@@ -9,16 +10,22 @@ import android.util.Log
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.snackbar.Snackbar
-import com.google.gson.*
+import com.google.gson.Gson
+import jp.shiguredo.sora.quickstart.databinding.ActivityMainBinding
 import jp.shiguredo.sora.sdk.camera.CameraCapturerFactory
 import jp.shiguredo.sora.sdk.channel.SoraMediaChannel
 import jp.shiguredo.sora.sdk.channel.option.SoraMediaOption
 import jp.shiguredo.sora.sdk.channel.signaling.message.PushMessage
 import jp.shiguredo.sora.sdk.error.SoraErrorReason
 import jp.shiguredo.sora.sdk.util.SoraLogger
-import kotlinx.android.synthetic.main.activity_main.*
-import org.webrtc.*
-import permissions.dispatcher.*
+import org.webrtc.CameraVideoCapturer
+import org.webrtc.EglBase
+import org.webrtc.MediaStream
+import permissions.dispatcher.NeedsPermission
+import permissions.dispatcher.OnPermissionDenied
+import permissions.dispatcher.OnShowRationale
+import permissions.dispatcher.PermissionRequest
+import permissions.dispatcher.RuntimePermissions
 
 @RuntimePermissions
 class MainActivity : AppCompatActivity() {
@@ -32,25 +39,28 @@ class MainActivity : AppCompatActivity() {
 
     private var audioManager: AudioManager? = null
 
+    private lateinit var binding: ActivityMainBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         SoraLogger.enabled = true
 
-        setContentView(R.layout.activity_main)
-        startButton.setOnClickListener {
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        binding.startButton.setOnClickListener {
             disableStartButton()
             startWithPermissionCheck()
         }
-        stopButton.setOnClickListener {
+        binding.stopButton.setOnClickListener {
             close()
             disableStopButton()
         }
 
         egl = EglBase.create()
         val eglContext = egl!!.eglBaseContext
-        localRenderer?.init(eglContext, null)
-        remoteRenderer?.init(eglContext, null)
+        binding.localRenderer?.init(eglContext, null)
+        binding.remoteRenderer?.init(eglContext, null)
         disableStopButton()
 
         audioManager = applicationContext.getSystemService(Context.AUDIO_SERVICE) as AudioManager
@@ -64,6 +74,8 @@ class MainActivity : AppCompatActivity() {
         this.volumeControlStream = AudioManager.STREAM_VOICE_CALL
     }
 
+    // AudioManager.MODE_INVALID が使われているため lint でエラーが出るので一時的に抑制しておく
+    @SuppressLint("WrongConstant")
     override fun onDestroy() {
         Log.d(TAG, "onDestroy")
         super.onDestroy()
@@ -104,7 +116,7 @@ class MainActivity : AppCompatActivity() {
                 if (ms.videoTracks.size > 0) {
                     val track = ms.videoTracks[0]
                     track.setEnabled(true)
-                    track.addSink(this@MainActivity.remoteRenderer)
+                    track.addSink(this@MainActivity.binding.remoteRenderer)
                 }
             }
         }
@@ -115,18 +127,18 @@ class MainActivity : AppCompatActivity() {
                 if (ms.videoTracks.size > 0) {
                     val track = ms.videoTracks[0]
                     track.setEnabled(true)
-                    track.addSink(this@MainActivity.localRenderer)
+                    track.addSink(this@MainActivity.binding.localRenderer)
                     capturer?.startCapture(400, 400, 30)
                 }
             }
         }
 
         override fun onPushMessage(mediaChannel: SoraMediaChannel, push: PushMessage) {
-            Log.d(TAG, "onPushMessage: push=${push}")
+            Log.d(TAG, "onPushMessage: push=$push")
             val data = push.data
-            if(data is Map<*, *>) {
+            if (data is Map<*, *>) {
                 data.forEach { (key, value) ->
-                    Log.d(TAG, "received push data: ${key}=${value}")
+                    Log.d(TAG, "received push data: $key=$value")
                 }
             }
         }
@@ -150,12 +162,13 @@ class MainActivity : AppCompatActivity() {
         }
 
         mediaChannel = SoraMediaChannel(
-                context           = this,
-                signalingEndpoint = BuildConfig.SIGNALING_ENDPOINT,
-                channelId         = BuildConfig.CHANNEL_ID,
-                signalingMetadata = Gson().fromJson(BuildConfig.SIGNALING_METADATA, Map::class.java),
-                mediaOption       = option,
-                listener          = channelListener)
+            context = this,
+            signalingEndpoint = BuildConfig.SIGNALING_ENDPOINT,
+            channelId = BuildConfig.CHANNEL_ID,
+            signalingMetadata = Gson().fromJson(BuildConfig.SIGNALING_METADATA, Map::class.java),
+            mediaOption = option,
+            listener = channelListener
+        )
         mediaChannel!!.connect()
     }
 
@@ -169,25 +182,25 @@ class MainActivity : AppCompatActivity() {
         capturer?.stopCapture()
         capturer = null
 
-        localRenderer?.release()
-        remoteRenderer?.release()
+        binding.localRenderer?.release()
+        binding.remoteRenderer?.release()
 
         egl?.release()
         egl = null
     }
 
     private fun disableStartButton() {
-        stopButton.isEnabled = true
-        stopButton.setBackgroundColor(Color.parseColor("#F06292"))
-        startButton.isEnabled = false
-        startButton.setBackgroundColor(Color.parseColor("#CCCCCC"))
+        binding.stopButton.isEnabled = true
+        binding.stopButton.setBackgroundColor(Color.parseColor("#F06292"))
+        binding.startButton.isEnabled = false
+        binding.startButton.setBackgroundColor(Color.parseColor("#CCCCCC"))
     }
 
     private fun disableStopButton() {
-        stopButton.isEnabled = false
-        stopButton.setBackgroundColor(Color.parseColor("#CCCCCC"))
-        startButton.isEnabled = true
-        startButton.setBackgroundColor(Color.parseColor("#F06292"))
+        binding.stopButton.isEnabled = false
+        binding.stopButton.setBackgroundColor(Color.parseColor("#CCCCCC"))
+        binding.startButton.isEnabled = true
+        binding.startButton.setBackgroundColor(Color.parseColor("#F06292"))
     }
 
     // -- PermissionDispatcher --
@@ -201,25 +214,28 @@ class MainActivity : AppCompatActivity() {
     fun showRationaleForCameraAndAudio(request: PermissionRequest) {
         Log.d(TAG, "showRationaleForCameraAndAudio")
         showRationaleDialog(
-                "ビデオチャットを利用するには、カメラとマイクの使用許可が必要です", request)
+            "ビデオチャットを利用するには、カメラとマイクの使用許可が必要です", request
+        )
     }
 
     @OnPermissionDenied(value = [Manifest.permission.CAMERA, Manifest.permission.RECORD_AUDIO])
     fun onCameraAndAudioDenied() {
         Log.d(TAG, "onCameraAndAudioDenied")
-        Snackbar.make(rootLayout,
-                "ビデオチャットを利用するには、カメラとマイクの使用を許可してください",
-                Snackbar.LENGTH_LONG)
-                .setAction("OK") { }
-                .show()
+        Snackbar.make(
+            binding.rootLayout,
+            "ビデオチャットを利用するには、カメラとマイクの使用を許可してください",
+            Snackbar.LENGTH_LONG
+        )
+            .setAction("OK") { }
+            .show()
     }
 
     private fun showRationaleDialog(message: String, request: PermissionRequest) {
         AlertDialog.Builder(this)
-                .setPositiveButton(getString(R.string.permission_button_positive)) { _, _ -> request.proceed() }
-                .setNegativeButton(getString(R.string.permission_button_negative)) { _, _ -> request.cancel() }
-                .setCancelable(false)
-                .setMessage(message)
-                .show()
+            .setPositiveButton(getString(R.string.permission_button_positive)) { _, _ -> request.proceed() }
+            .setNegativeButton(getString(R.string.permission_button_negative)) { _, _ -> request.cancel() }
+            .setCancelable(false)
+            .setMessage(message)
+            .show()
     }
 }